requirements.txt

telethon
python-dotenv
